day06 1. 修改了查询菜品的参数类型 Integer -> Long 2. 修改了菜品分页查询形参的参数类型DishPageQueryDTO -> Dish 3. 修改了菜品分页查询的sql语句  添加查询categoryId，description 字段 4. 导入了 用户端浏览商品代码 4.1 新建了用户端的菜品、套餐、分类的控制层

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -1,5 +1,6 @@
 package com.sky.controller.admin;
 
+import com.sky.constant.StatusConstant;
 import com.sky.dto.DishDTO;
 import com.sky.dto.DishPageQueryDTO;
 import com.sky.entity.Dish;
@@ -63,8 +64,10 @@ public class DishController {
      *  1. 查询菜品列表
      */
     @GetMapping("/list")
-    public Result<List<Dish>> selectDish(Integer categoryId){
+    public Result<List<Dish>> selectDish(Long categoryId){
         List<Dish> list = dishService.selectDishByCategoryId(categoryId);
         return Result.success(list);
     }
+
+
 }
